Airspark improvements, main menu framework, and a lot more
Changed Samus's hurtbox from a StaticBody2D to a KinematicBody2D, because there is a [Godot engine bug](https://github.com/godotengine/godot/issues/17238) which makes Area2Ds unable to detect StaticBody2Ds

Applied the ExArea2D node to several existing scenes
Various bugfixes and optimisations

Added the SamusCannonPosition scene to assist with the WeaponManager and add additional functionality
Used the SamusCannonPosition to enable carrying a chargebeam during an airspark (part of #50)

Added an animation to the SaveStation for when Samus spawns into the room when starting the game (part of #35)
Reorganised the global game settings by renaming the "window" group to "visuals", and moving several settings to more appropriate groups
Added the main_menu_scene_preview setting, which will determine whether the most recent room of each savefile will be loaded and displayed when selecting a savefile in the mainmenu (part of #35)

The Enums singleton now manages setting and getting node groups
Fixed the collision of doors from stopping Samus's momenum when she enters a door at high speed (such as when shinesparking)
Added the can_airspark_while_grounded physics value. It's set to off by default (part of #50).