Added many sounds (including for landing and stepping), improved and expanded difficulty selection
The SamusState class now automatically loads the sounds for its state ID, like it does with animations
Added and completed sound effects for (part of #8):
 - Samus:
	 - Beam firing
	 - Missile firing and exploding
	 - Taking damage
	 - Spiderball activation and crawl
	 - Speedbooster, speedbooster brake, shinespark charge and shinespark flight
	 - Powergrip ledge grab and climb
	 - Bomb and powerbomb placement and explosion
 - Room earthquakes

Per-state gravity and fall_speed_cap overrides in Samus's physics data must now end with "_override"
When landing after a jump, Samus plays the crouch/uncrouch animation if she is standing still
Renamed the Shortcut singleton to InputManager
In the settings_information data json, settings options can now have the "required_settings" key. The option will only be shown in the interface if the value of each option in "required_settings" matches the value of the option in the actual settings.
Fixed the ButtonIcon script to allow the icon to be set before the node enters the scene tree, as this broke the ButtonPromptNotification display

Samus now plays landing and step sounds based on the surface she is standing on (part of #8):
 - Added the GroundTypeCollisionShape2D class, which is just a Node2D containing data on the different ground types and an export var to select one
 - When Samus lands on the floor, the GroundTypeCollisionShape2D she is standing on is found and stored
 - The GroundTypeCollisionShape2D manages and plays the sounds for landing and stepping

Improved the difficulty selection dialog and MultipleSelectionPrompt (part of #35):
 - Added difficulty descriptions, which are displayed when selecting them
 - Individual difficulty data values can now be hidden and locked (eg. unlock a harder difficulty upon beating the game on normal)
 - Non-hidden data values are now displayed on the selection screen as a title and a value formatted by type
 - It looks a lot less horrible
Added a very hard difficulty and a one-hit KO difficulty (both Samus and enemies die in one hit)
All existing difficulty options now apply correctly in-game (incoming/outgoing_damage_multiplier, ohko_samus, and ohko_enemies)